Rename BellMessage component and add 'client' postfix to Bell component
