+++ src/app/planet/[communityId]/id-card/[idCardId]/components/IdCardDetail/IdCardDetail.tsx
@@ -4,7 +4,7 @@ import { Suspense } from 'react';
 
 import { useGetIdCardDetail } from '~/api/domain/idCard.api';
 import { useGetUserInfo } from '~/api/domain/user.api';
-import { Bell } from '~/app/planet/[communityId]/id-card/[idCardId]/components/Bell/Bell';
+import { Bell } from '~/app/planet/[communityId]/id-card/[idCardId]/components/Bell/Bell.client';
 import RetryErrorBoundary from '~/components/ErrorBoundary/RetryErrorBoundary.client';
 import { TopNavigation } from '~/components/TopNavigation';
 import { Intro, KeywordContentCard } from '~/modules/IdCardDetail';
